+++ tests/testthat/test-roster.R
@@ -3,7 +3,7 @@ test_that("All current team rosters as list", {
   expect_type(r, "list")
   expect_length(r, 4)
   expect_s3_class(r[[1]], "data.frame")
-  expect_length(r[[1]], 15)
+  expect_gt(length(r[[1]]), 10)
   expect_s3_class(r[[1]]$lineupSlot, "factor")
 })